+++ source/Application/Model/Category.php
@@ -7,10 +7,9 @@
 
 namespace OxidEsales\EshopCommunity\Application\Model;
 
-use oxDb;
+use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\Eshop\Core\TableViewNameGenerator;
-use oxRegistry;
-use oxField;
+use OxidEsales\EshopCommunity\Application\Controller\FrontendController;
 
 /**
  * Category manager.
@@ -1134,9 +1133,13 @@ class Category extends \OxidEsales\Eshop\Core\Model\MultiLanguageModel implement
     public function getLongDesc()
     {
         if (isset($this->oxcategories__oxlongdesc) && $this->oxcategories__oxlongdesc instanceof \OxidEsales\Eshop\Core\Field) {
-            /** @var \OxidEsales\Eshop\Core\UtilsView $oUtilsView */
-            $oUtilsView = \OxidEsales\Eshop\Core\Registry::getUtilsView();
-            return $oUtilsView->parseThroughSmarty($this->oxcategories__oxlongdesc->getRawValue(), $this->getId() . $this->getLanguage(), null, true);
+            $activeView = oxNew(FrontendController::class);
+            $activeView->addGlobalParams();
+            $utilsView = Registry::getUtilsView();
+            return $utilsView->getRenderedContent(
+                $this->oxcategories__oxlongdesc->getRawValue(),
+                $activeView->getViewData(),
+                $this->getId() . $this->getLanguage());
         }
     }